Simplification Eq
The former proof of eq what ill-written.
The code_eq_arg was actually doing nothing.
I reformulated the def of FR_eq in order to see it more easly.

Finally the proof of IsFun was giving an element than proving that it is the only one.
I have given a proof by transforming the space to the instance of id type of {x : RA x y}.
This proof is clearer and quicker but not closer fron the vect one.