Change in vectF
To do the proof of FRvect_retr, I had the issue that the condition e : S m' = m was after the one on v.
However, the proof on v was not refl, hence I was left to prove transport _ p eq_rel = eq_refl.
Hence, I have changed the order to solve this issue.

I have one issue left, for some reason Coq is saying that there is unfocused goal
but I don't understand where.